feat: remove barrel files
