+++ components/HomeSections/Bundeswehr/index.tsx
