--- components/HomeSections/Darkweb/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Darkweb';
