--- components/HomeSections/DhiGrasDevelopment/index.ts
@@ -1,1 +1,0 @@
-export { default } from './DhiGrasDevelopment';
