--- components/HomeSections/Hero/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Hero';
