--- components/HomeSections/Reitzenstein/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Reitzenstein';
